Move file structure around
Required for testing to actually work and not use an address every single time
